Fix compilation errors on VC 7

+++ src/bios.cpp
@@ -1108,7 +1108,7 @@ void BIOS_Sqrt()
         VCOUNT);
   }
 #endif
-  reg[0].I = (u32)sqrt(reg[0].I);
+  reg[0].I = (u32)sqrt((double)reg[0].I);
 #ifdef DEV_VERSION
   if(systemVerbose & VERBOSE_SWI) {
     log("Sqrt: return=%08x\n",
@@ -1130,7 +1130,7 @@ void BIOS_MidiKey2Freq()
   int freq = CPUReadMemory(reg[0].I+4);
   double tmp;
   tmp = ((double)(180 - reg[1].I)) - ((double)reg[2].I / 256.f);
-  tmp = pow(2, tmp / 12.f);
+  tmp = pow((double)2.f, tmp / 12.f);
   reg[0].I = (int)((double)freq / tmp);
 
 #ifdef DEV_VERSION
